+++ HISTORY.rst
@@ -1,7 +1,10 @@
 0.2 (Unreleased)
 ----------------
 
-* 
+* Added an archive view (GitHub issue #10)
+* Added tags field to NewsItem model for admin integration (GitHub issue #7)
+* Added tags field to NewsItem administration (GitHub issue #7)
+* Added south migration for the tags field (GitHub issue #7)
 
 0.1 (2011/07/15)
 ----------------